requirements.txt
# Azure OpenAI
openai>=1.0.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Machine Learning
xgboost>=2.0.0

# Utilities
pathlib